requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.3
frozenlist==1.3.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
pyee==9.0.4
python-dotenv==0.20.0
slack-bolt==1.14.3
slack-sdk==3.18.1
slackclient==2.9.4
slackeventsapi==3.0.1
typing_extensions==4.3.0
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.1
